backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
bcrypt
python-jose[cryptography]
fastapi[all]
jinja2